Shapable added and CommonCheckBox.kt name changed to CommonStateButton.kt
