+++ app/src/main/res/layout-land/activity_private_player.xml
@@ -25,7 +25,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent">
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnCamera"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -35,7 +35,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-         <com.trx.consumer.common.CommonCheckBox
+         <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnMicrophone"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -45,7 +45,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btnCamera" />
 
-         <com.trx.consumer.common.CommonCheckBox
+         <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnClock"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -55,7 +55,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/btnMicrophone" />
 
-         <com.trx.consumer.common.CommonCheckBox
+         <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnShare"
             android:layout_width="45dp"
             android:layout_height="45dp"
