+++ app/src/main/res/layout/activity_group_player.xml
@@ -20,7 +20,7 @@
         android:layout_height="75dp"
         app:layout_constraintBottom_toBottomOf="parent">
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnCamera"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -30,7 +30,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"/>
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnMicrophone"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -40,7 +40,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"/>
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnClock"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -50,7 +50,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"/>
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnShare"
             android:layout_width="45dp"
             android:layout_height="45dp"
