+++ app/src/main/res/layout/activity_private_player.xml
@@ -20,17 +20,19 @@
         android:layout_height="75dp"
         app:layout_constraintBottom_toBottomOf="parent">
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnCamera"
             android:layout_width="45dp"
             android:layout_height="45dp"
+            app:cornerRadius="30dp"
+            android:backgroundTint="@color/blue"
             android:layout_marginStart="24dp"
             android:button="@drawable/ic_tab_camera"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"/>
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnMicrophone"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -40,7 +42,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"/>
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnClock"
             android:layout_width="45dp"
             android:layout_height="45dp"
@@ -50,7 +52,7 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"/>
 
-        <com.trx.consumer.common.CommonCheckBox
+        <com.trx.consumer.common.CommonStateButton
             android:id="@+id/btnShare"
             android:layout_width="45dp"
             android:layout_height="45dp"
